move ruleset into new draftfast module
